pyproject.toml
[tool.poetry]
name = "discord-jab"
version = "0.1.0"
description = "Just another bot. Exploring python discord bot usage"
authors = ["James Green <Greenjam94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
"discord.py" = "^1.7.3"
async-rediscache = "^0.2.0"
aiodns = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
